requirements.txt

asgiref==3.6.0
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django>=4.2
django-cors-headers==3.14.0
django-rest-swagger==2.2.0
djangorestframework>=3.14.0
environ==1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
razorpay==1.3.0
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
gunicorn==20.1.0
